+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- No longer send a response body when one has not been supplied.
+
 ## 1.0 - 2017-12-02
 ### Changed
 - Require PHP 7.0+